reference/scenarios/google-genai/pyproject.toml
[project]
name = "google-genai-reference-test"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.72.0",
    "genai-reference-shared",
]

[tool.uv.sources]
genai-reference-shared = { path = "../../shared", editable = true }

[tool.uv]
package = false
